requirements.txt
fastapi==0.115.6
Flask==3.1.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
newsapi-python==0.2.7
numpy==1.26.4
openai==1.57.0
psycopg2-binary==2.9.9
pyarrow==18.1.0
pydantic==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
python-multipart==0.0.20
requests==2.32.3
requests-toolbelt==1.0.0
sentence-transformers==3.4.1
tiktoken==0.8.0
urllib3==2.2.3
uvicorn==0.32.1